# Conflicts:
#	src/main/java/core/AbstractGameState.java
#	src/main/java/players/mcts/SingleTreeNode.java

+++ src/main/java/games/coltexpress/ColtExpressGameState.java
@@ -244,7 +244,7 @@ public class ColtExpressGameState extends AbstractGameStateWithTurnOrder impleme
         for (ColtExpressCard card : playerHandCards.get(playerId).getComponents())
             if (card.cardType == ColtExpressCard.CardType.Bullet)
                 bulletsTaken++;
-        return (double) -bulletsTaken;
+        return -bulletsTaken;
     }
 
     @Override
