+++ src/main/java/games/sushigo/SGGameState.java
@@ -191,7 +191,7 @@ public class SGGameState extends AbstractGameState {
     @Override
     public double getTiebreak(int playerId) {
         // Tie-break is number of puddings
-        return (double) playedCardTypes[playerId].get(SGCard.SGCardType.Pudding).getValue();
+        return playedCardTypes[playerId].get(SGCard.SGCardType.Pudding).getValue();
     }
 
     @Override
